소셜 유저 토큰 검증
